refactor: omit empty client-secret from config file when logging in
This commit does not store the client secret in the configuration
file if it is empty when logging in. This is the case in a normal
login scenario when using the OpenID Connect Authorization Code
Flow with PKCE.

As stated in the previous commit this application is considered a
public client and does not require the user to keep the client
secret confidential. It is still possible to manually set the
client secret field in the configuration file, but it will be
removed when logging out.