refactor!: remove hardcoded client secret
A client secret allows the authorization server to determine the
identity of the client.

This is recommended for confidential clients and PKCE is an
additional layer of security.

Since this application is considered a public client it does not
make sense to use a default client secret that is hardcoded in the
application.

The reason it was there in the first place was lack of knowledge.