refactor: remove non-standard scope
This commit removes a scope that is not defined by the OpenID Connect
specification and DOES NOT WORK with the Keycloak authorization server.

The specification states that scope values used that are not
understood by an application should be IGNORED, but apparently not all
authorization servers adheres to the spec.

Ref: https://openid.net/specs/openid-connect-basic-1_0.html#Scopes