feat: add PKCE to the Authorization Code Flow
This commit implements Proof Key for Code Exchange (PKCE) for the
Authorization Code Flow.

See https://oauth.net/2/pkce/ and https://tools.ietf.org/html/rfc7636.

PKCE ensures that the client application exchanging the authorization
code is the same client application that requested it and that the
authorization code was not stolen and injected in a different session.